+++ vcxproj2cmake.Tests/ConverterTests/IncludeHeadersTests.cs
@@ -40,7 +40,7 @@ public partial class ConverterTests
                 """)));
 
             var converter = new Converter(fileSystem, new InMemoryLogger());
-            converter.Convert(projectFiles: [new FileInfo(@"Project.vcxproj")], dryRun: false, includeHeaders: false);
+            converter.Convert(projectFiles: [new FileInfo(@"Project.vcxproj")], includeHeaders: false);
 
             var cmake = fileSystem.GetFile(@"CMakeLists.txt").TextContents;
             Assert.Contains("""                
@@ -68,13 +68,74 @@ public partial class ConverterTests
                 """)));
 
             var converter = new Converter(fileSystem, new InMemoryLogger());
-            converter.Convert(projectFiles: [new FileInfo(@"Project.vcxproj")], dryRun: false, includeHeaders: true);
+            converter.Convert(projectFiles: [new FileInfo(@"Project.vcxproj")], includeHeaders: true);
 
             var cmake = fileSystem.GetFile(@"CMakeLists.txt").TextContents;
             Assert.Contains("""                
                 target_sources(Project
                     PRIVATE
                         include/foo.h
+                        src/main.cpp
+                )
+                """, cmake);
+        }
+
+        [Fact]
+        public void Given_ProjectWithQtMocHeaders_When_IncludeHeadersIsTrue_Then_TargetSourcesListsQtMocHeaders()
+        {
+            var fileSystem = new MockFileSystem();
+            fileSystem.Directory.SetCurrentDirectory(Environment.CurrentDirectory);
+
+            fileSystem.AddFile(@"Project.vcxproj", new(CreateProject("""
+                <ItemGroup>
+                    <ClCompile Include="src\main.cpp" />
+                </ItemGroup>
+                <ItemGroup>
+                    <QtMoc Include="include\moc.H" />
+                    <QtMoc Include="include\moc.HpP" />
+                    <QtMoc Include="include\moc.HxX" />
+                    <QtMoc Include="include\moc.H++" />
+                    <QtMoc Include="include\moc.Hh" />
+                </ItemGroup>
+                """)));
+
+            var converter = new Converter(fileSystem, new InMemoryLogger());
+            converter.Convert(projectFiles: [new FileInfo(@"Project.vcxproj")], includeHeaders: true);
+
+            var cmake = fileSystem.GetFile(@"CMakeLists.txt").TextContents;
+            Assert.Contains("""                
+                target_sources(Project
+                    PRIVATE
+                        include/moc.H
+                        include/moc.H++
+                        include/moc.Hh
+                        include/moc.HpP
+                        include/moc.HxX
+                        src/main.cpp
+                )
+                """, cmake);
+        }
+
+        [Fact]
+        public void Given_ProjectWithQtMocSourceFiles_When_IncludeHeadersIsTrue_Then_TargetSourcesDoesNOtListQtMocSources()
+        {
+            var fileSystem = new MockFileSystem();
+            fileSystem.Directory.SetCurrentDirectory(Environment.CurrentDirectory);
+
+            fileSystem.AddFile(@"Project.vcxproj", new(CreateProject("""
+                <ItemGroup>
+                    <ClCompile Include="src\main.cpp" />
+                    <QtMoc Include="src\moc.cpp" />
+                </ItemGroup>
+                """)));
+
+            var converter = new Converter(fileSystem, new InMemoryLogger());
+            converter.Convert(projectFiles: [new FileInfo(@"Project.vcxproj")], includeHeaders: true);
+
+            var cmake = fileSystem.GetFile(@"CMakeLists.txt").TextContents;
+            Assert.Contains("""                
+                target_sources(Project
+                    PRIVATE
                         src/main.cpp
                 )
                 """, cmake);